Fix: merge workflow pkg convert to pkg util
