--- pkg/stdlib/pkgs/convert.cue
@@ -1,8 +1,0 @@
-#String: {
-	#do:       "string"
-	#provider: "convert"
-
-	bt:   bytes
-	str?: string
-	...
-}
